feat: update and add new images for various gift projects
Update existing images in the 202502_Valentine_gifts and
202505_Tool_Organizer directories to reflect the latest designs
and improvements. Add new images for the 202505_Heidi_BD_Gift,
202506_Birthday_Gift_Heidi, 202509_Birthday_Gift_Evan-M, and
Boda_lego_car_garage projects to provide visual previews of the
gifts and structures.

These changes enhance the visual documentation of the projects,
making it easier for users to understand the design and appearance
of each gift. The addition of preview images in README files
improves the user experience by providing immediate visual context.